+++ crates/backend-uzu/src/config/rope/longrope.rs
@@ -4,6 +4,6 @@ use proc_macros::uzu_config;
 pub struct LongRoPEConfig {
     pub short_factor: Box<[f32]>,
     pub long_factor: Box<[f32]>,
-    pub original_context_length: usize,
+    pub original_context_length: u32,
     pub scaling_factor: f32,
 }
